Added buy_sell function which will handle the displaying of all avalible items that the user can buy or sell.
Moved the inventory function to check_inventory which is a more
proper name and now a method of  player.

Updated inventory class so when user goes to the room they can
choose to drop, add, and see their inventory.

Added a few todo comments an removed old todo comments to
keep project flow.