Created the buy_sell, and the inventory class.
Created the  drop method and Updated
the add_to_inventory method.

The buy_sell class was created to allow the user to get
new items, sell their items, and see a full list of the
items in the game.

The inventory class was created to call the add_to_inventory methd and
the drop method so the user could update and see their inventory as the
please.

Lastly, The add_to_inventory method and it's associated methods where
created to fill the roles of the inventory class.