Updated the find_item function to allow finding of weapons just in the list_of_items for the game.
Updated the check_inventory method to list out inventory values
in a table like fashion on the command line.

Created and finished the buying function which allows users
to pick and purchase items in the menu.

Updated add_to_player_inventory method to accomidate for all
types of items being added to inventory.