I updated the gitignore to also ignore .pyc files to decrease clutter.
I created and updated the RULES.md file which gives an intro to the game 
and details for the user.

I created two more rooms SafePlace, and Normandy to continue the game.

I moved all docs to the docs directory to decrease clutter files.

Lastly, I updated the add_to_inventory method which now checks and then 
adds items to players inventory.